requirements.txt
Flask
gunicorn
mysql-connector-python
flask-cors
psycopg2-binary